Extract openers.py from legacy.py
